requirements.txt
anyio==3.7.1
APScheduler==3.6.3
cachetools==4.2.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
Deprecated==1.2.14
exceptiongroup==1.1.2
h11==0.14.0
httpcore==0.16.0
httpx==0.23.1
idna==3.4
jproperties==2.1.1
pycparser==2.21
PyGithub==1.59.1
PyJWT==2.8.0
PyNaCl==1.5.0
python-telegram-bot==20.0
pytz==2023.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
tornado==6.2
typing_extensions==4.7.1
tzlocal==5.0.1
urllib3==1.26.6
wrapt==1.15.0
